backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
anthropic==0.7.0
pytest==7.4.0
black==23.9.1
isort==5.12.0
flask-cors==4.0.0
reportlab==4.0.4
